Split utils into separate modules
*  Separates utils.py into multiple modules.
*  Updates references to refactored utils.
