+++ packit/patches.py
@@ -35,7 +35,8 @@ import yaml
 from packit.constants import DATETIME_FORMAT
 from packit.git_utils import get_metadata_from_message
 from packit.local_project import LocalProject
-from packit.utils import is_a_git_ref, run_command
+from packit.utils.repo import is_a_git_ref
+from packit.utils.commands import run_command
 
 logger = logging.getLogger(__name__)
