+++ packit/utils/version.py
@@ -1,6 +1,6 @@
 # MIT License
 #
-# Copyright (c) 2019 Red Hat, Inc.
+# Copyright (c) 2018-2019 Red Hat, Inc.
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -20,26 +20,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from packit.utils.commands import cwd
-
-from tests.spellbook import call_packit
+from pkg_resources import get_distribution, DistributionNotFound
 
 
-def test_generate_pass(upstream_without_config):
-    with cwd(upstream_without_config):
-        assert not (upstream_without_config / ".packit.yaml").is_file()
-
-        # This test requires packit on pythonpath
-        result = call_packit(parameters=["generate"])
-
-        assert result.exit_code == 0
-
-        assert (upstream_without_config / ".packit.yaml").is_file()
-
-
-def test_generate_fail(cwd_upstream_or_distgit):
-    result = call_packit(
-        parameters=["generate"], working_dir=str(cwd_upstream_or_distgit)
-    )
-
-    assert result.exit_code == 2  # packit config already exists --force needed
+def get_packit_version() -> str:
+    try:
+        return get_distribution("packitos").version
+    except DistributionNotFound:
+        return "NOT_INSTALLED"
